pyproject.toml
[project]
name = "itsascout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.14.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-object-actions>=5.0.0",
    "django-rq>=3.2.2",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.0",
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "inertia-django>=1.2.0",
    "logfire>=3.24.0",
    "loguru>=0.7.3",
    "polars>=1.31.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[openai]>=0.4.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sh>=2.2.2",
    "tqdm>=4.67.1",
    "w3lib>=2.4.0",
    "wafw00f>=2.3.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "scrapegrape.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["scrapegrape"]
